+++ lib/supply_chain.dart
@@ -9,7 +9,6 @@ library;
 export 'src/doc.dart';
 export 'src/graph.dart';
 export 'src/insert.dart';
-export 'src/insert_blue_print.dart';
 export 'src/inserts.dart';
 export 'src/keys.dart';
 export 'src/node_blue_print.dart';